- moved kafka connect related endpoints to separate tag/controller
