+++ kafka-ui-api/src/main/java/com/provectus/kafka/ui/rest/MetricsRestController.java
@@ -3,7 +3,6 @@ package com.provectus.kafka.ui.rest;
 import com.provectus.kafka.ui.api.ApiClustersApi;
 import com.provectus.kafka.ui.cluster.model.ConsumerPosition;
 import com.provectus.kafka.ui.cluster.service.ClusterService;
-import com.provectus.kafka.ui.cluster.service.KafkaConnectService;
 import com.provectus.kafka.ui.cluster.service.SchemaRegistryService;
 import com.provectus.kafka.ui.model.*;
 import lombok.RequiredArgsConstructor;
@@ -19,7 +18,6 @@ import reactor.core.publisher.Mono;
 import javax.validation.Valid;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.function.Function;
 
 @RestController
@@ -29,7 +27,6 @@ public class MetricsRestController implements ApiClustersApi {
 
     private final ClusterService clusterService;
     private final SchemaRegistryService schemaRegistryService;
-    private final KafkaConnectService kafkaConnectService;
 
     @Override
     public Mono<ResponseEntity<Flux<Cluster>>> getClusters(ServerWebExchange exchange) {
@@ -187,96 +184,6 @@ public class MetricsRestController implements ApiClustersApi {
     public Mono<ResponseEntity<Void>> updateSchemaCompatibilityLevel(String clusterName, String subject, @Valid Mono<CompatibilityLevel> compatibilityLevel, ServerWebExchange exchange) {
         log.info("Updating schema compatibility for subject: {}", subject);
         return schemaRegistryService.updateSchemaCompatibility(clusterName, subject, compatibilityLevel)
-                .map(ResponseEntity::ok);
-    }
-
-    @Override
-    public Mono<ResponseEntity<Flux<String>>> getConnectors(String clusterName, ServerWebExchange exchange) {
-        Flux<String> connectors = kafkaConnectService.getConnectors(clusterName);
-        return Mono.just(ResponseEntity.ok(connectors));
-    }
-
-    @Override
-    public Mono<ResponseEntity<Connector>> createConnector(String clusterName, @Valid Mono<NewConnector> connector, ServerWebExchange exchange) {
-        return kafkaConnectService.createConnector(clusterName, connector)
-                .map(ResponseEntity::ok);
-    }
-
-    @Override
-    public Mono<ResponseEntity<Connector>> getConnector(String clusterName, String connectorName, ServerWebExchange exchange) {
-        return kafkaConnectService.getConnector(clusterName, connectorName)
-                .map(ResponseEntity::ok);
-    }
-
-    @Override
-    public Mono<ResponseEntity<Void>> deleteConnector(String clusterName, String connectorName, ServerWebExchange exchange) {
-        // todo return 409
-        return kafkaConnectService.deleteConnector(clusterName, connectorName)
-                .map(v -> ResponseEntity.noContent().build());
-    }
-
-    @Override
-    public Mono<ResponseEntity<Map<String, Object>>> getConnectorConfig(String clusterName, String connectorName, ServerWebExchange exchange) {
-        return kafkaConnectService.getConnectorConfig(clusterName, connectorName)
-                .map(ResponseEntity::ok);
-    }
-
-    @Override
-    public Mono<ResponseEntity<Connector>> setConnectorConfig(String clusterName, String connectorName, @Valid Mono<Object> requestBody, ServerWebExchange exchange) {
-        return kafkaConnectService.setConnectorConfig(clusterName, connectorName, requestBody)
-                .map(ResponseEntity::ok);
-    }
-
-    @Override
-    public Mono<ResponseEntity<ConnectorStatus>> getConnectorStatus(String clusterName, String connectorName, ServerWebExchange exchange) {
-        return kafkaConnectService.getConnectorStatus(clusterName, connectorName)
-                .map(ResponseEntity::ok);
-    }
-
-    @Override
-    public Mono<ResponseEntity<Void>> restartConnector(String clusterName, String connectorName, ServerWebExchange exchange) {
-        return kafkaConnectService.restartConnector(clusterName, connectorName)
-                .map(ResponseEntity::ok);
-    }
-
-    @Override
-    public Mono<ResponseEntity<Void>> pauseConnector(String clusterName, String connectorName, ServerWebExchange exchange) {
-        return kafkaConnectService.pauseConnector(clusterName, connectorName)
-                .map(ResponseEntity::ok);
-    }
-
-    @Override
-    public Mono<ResponseEntity<Void>> resumeConnector(String clusterName, String connectorName, ServerWebExchange exchange) {
-        return kafkaConnectService.resumeConnector(clusterName, connectorName)
-                .map(ResponseEntity::ok);
-    }
-
-    @Override
-    public Mono<ResponseEntity<Flux<ConnectorTask>>> getConnectorTasks(String clusterName, String connectorName, ServerWebExchange exchange) {
-        return Mono.just(ResponseEntity.ok(kafkaConnectService.getConnectorTasks(clusterName, connectorName)));
-    }
-
-    @Override
-    public Mono<ResponseEntity<TaskStatus>> getConnectorTaskStatus(String clusterName, String connectorName, Integer taskId, ServerWebExchange exchange) {
-        return kafkaConnectService.getConnectorTaskStatus(clusterName, connectorName, taskId)
-                .map(ResponseEntity::ok);
-    }
-
-    @Override
-    public Mono<ResponseEntity<Void>> restartConnectorTask(String clusterName, String connectorName, Integer taskId, ServerWebExchange exchange) {
-        return kafkaConnectService.restartConnectorTask(clusterName, connectorName, taskId)
-                .map(ResponseEntity::ok);
-    }
-
-    @Override
-    public Mono<ResponseEntity<Flux<ConnectorPlugin>>> getConnectorPlugins(String clusterName, ServerWebExchange exchange) {
-        return kafkaConnectService.getConnectorPlugins(clusterName)
-                .map(ResponseEntity::ok);
-    }
-
-    @Override
-    public Mono<ResponseEntity<ConnectorPluginConfigValidationResponse>> validateConnectorPluginConfig(String clusterName, String pluginName, @Valid Mono<Object> requestBody, ServerWebExchange exchange) {
-        return kafkaConnectService.validateConnectorPluginConfig(clusterName, pluginName, requestBody)
                 .map(ResponseEntity::ok);
     }
